--- Assets/KillAfterTime.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6c5d50534f45b634f9fc75369238da61
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
