--- Assets/MidiController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b380f0954487f124b9c7917d7986995d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
