+++ Assets/StreamingAssets/Midi.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0238b9646210e204f9a7bbdc4a1522be
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
